fix: wire up logger correctly and move RankingItem to lib
- chr_pipeline & svineflytning_pipeline: assign setup_pipeline_logger() return value to module logger via global declaration in setup_logging(), fixing silent log loss from logger/getLogger(__name__) mismatch
- Move RankingItem interface from src/types/ (reserved for .d.ts third-party types) to src/lib/rankings.ts per frontend/CLAUDE.md
- Update imports in RankingTable and IndividualRankingTable
